requirements.txt
aiohttp
requests
Pillow